STORAGE: Making Connection.get_all_buckets a standalone method.
See comments in #700 for context.
