+++ gcloud/storage/demo/demo.py
@@ -1,18 +1,3 @@
-# Copyright 2014 Google Inc. All rights reserved.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-# pragma NO COVER
 # Welcome to the gCloud Storage Demo! (hit enter)
 
 # We're going to walk through some of the basics...,
@@ -21,12 +6,13 @@
 # Let's start by importing the demo module and getting a connection:
 import time
 
+from gcloud import storage
 from gcloud.storage import demo
 
 connection = demo.get_connection()
 
 # OK, now let's look at all of the buckets...
-print(connection.get_all_buckets())  # This might take a second...
+print(list(storage.get_all_buckets(connection)))  # This might take a second...
 
 # Now let's create a new bucket...
 bucket_name = ("bucket-%s" % time.time()).replace(".", "")  # Get rid of dots.
@@ -35,7 +21,7 @@ bucket = connection.create_bucket(bucket_name)
 print(bucket)
 
 # Let's look at all of the buckets again...
-print(connection.get_all_buckets())
+print(list(storage.get_all_buckets(connection)))
 
 # How about we create a new blob inside this bucket.
 blob = bucket.new_blob("my-new-file.txt")
