+++ gcloud/storage/test_api.py
@@ -80,6 +80,102 @@ class Test_lookup_bucket(unittest2.TestCase):
         self._lookup_bucket_hit_helper(use_default=True)
 
 
+class Test_get_all_buckets(unittest2.TestCase):
+
+    def _callFUT(self, connection=None):
+        from gcloud.storage.api import get_all_buckets
+        return get_all_buckets(connection=connection)
+
+    def test_empty(self):
+        from gcloud.storage.connection import Connection
+        PROJECT = 'project'
+        conn = Connection(PROJECT)
+        URI = '/'.join([
+            conn.API_BASE_URL,
+            'storage',
+            conn.API_VERSION,
+            'b?project=%s' % PROJECT,
+        ])
+        http = conn._http = Http(
+            {'status': '200', 'content-type': 'application/json'},
+            '{}',
+        )
+        buckets = list(self._callFUT(conn))
+        self.assertEqual(len(buckets), 0)
+        self.assertEqual(http._called_with['method'], 'GET')
+        self.assertEqual(http._called_with['uri'], URI)
+
+    def _get_all_buckets_non_empty_helper(self, use_default=False):
+        from gcloud.storage._testing import _monkey_defaults
+        from gcloud.storage.connection import Connection
+        PROJECT = 'project'
+        BUCKET_NAME = 'bucket-name'
+        conn = Connection(PROJECT)
+        URI = '/'.join([
+            conn.API_BASE_URL,
+            'storage',
+            conn.API_VERSION,
+            'b?project=%s' % PROJECT,
+        ])
+        http = conn._http = Http(
+            {'status': '200', 'content-type': 'application/json'},
+            '{"items": [{"name": "%s"}]}' % BUCKET_NAME,
+        )
+
+        if use_default:
+            with _monkey_defaults(connection=conn):
+                buckets = list(self._callFUT())
+        else:
+            buckets = list(self._callFUT(conn))
+
+        self.assertEqual(len(buckets), 1)
+        self.assertEqual(buckets[0].name, BUCKET_NAME)
+        self.assertEqual(http._called_with['method'], 'GET')
+        self.assertEqual(http._called_with['uri'], URI)
+
+    def test_non_empty(self):
+        self._get_all_buckets_non_empty_helper(use_default=False)
+
+    def test_non_use_default(self):
+        self._get_all_buckets_non_empty_helper(use_default=True)
+
+
+class Test__BucketIterator(unittest2.TestCase):
+
+    def _getTargetClass(self):
+        from gcloud.storage.api import _BucketIterator
+        return _BucketIterator
+
+    def _makeOne(self, *args, **kw):
+        return self._getTargetClass()(*args, **kw)
+
+    def test_ctor(self):
+        connection = object()
+        iterator = self._makeOne(connection)
+        self.assertTrue(iterator.connection is connection)
+        self.assertEqual(iterator.path, '/b')
+        self.assertEqual(iterator.page_number, 0)
+        self.assertEqual(iterator.next_page_token, None)
+
+    def test_get_items_from_response_empty(self):
+        connection = object()
+        iterator = self._makeOne(connection)
+        self.assertEqual(list(iterator.get_items_from_response({})), [])
+
+    def test_get_items_from_response_non_empty(self):
+        from gcloud.storage.bucket import Bucket
+        BLOB_NAME = 'blob-name'
+        response = {'items': [{'name': BLOB_NAME}]}
+        connection = object()
+        iterator = self._makeOne(connection)
+        buckets = list(iterator.get_items_from_response(response))
+        self.assertEqual(len(buckets), 1)
+        bucket = buckets[0]
+        self.assertTrue(isinstance(bucket, Bucket))
+        self.assertTrue(bucket.connection is connection)
+        self.assertEqual(bucket.name, BLOB_NAME)
+
+
 class Http(object):
 
     _called_with = None
